Improve responsiveness for uploading files
