+++ src/routes/index.tsx
@@ -2,7 +2,6 @@ import { Navigate, createBrowserRouter } from "react-router-dom";
 import PageNotFound from "../components/page-not-found";
 import Layout from "../screens/layout";
 import MyDrive from "../screens/layout/my-drive";
-import Home from "../screens/Home";
 import Login from "../screens/auth/login";
 import ForgetPassword from "../screens/auth/forgot-password";
 import AuthGuard from "../components/guard";
@@ -23,7 +22,6 @@ export default createBrowserRouter([
         errorElement: <PageNotFound />,
         children: [
             {
-                path: "",
                 element: <Layout />,
                 children: [
                     {
@@ -41,11 +39,6 @@ export default createBrowserRouter([
                         element: <MyDrive />,
                         errorElement: <PageNotFound />
                     },
-                    {
-                        path: "test",
-                        element: <Home />,
-                        errorElement: <PageNotFound />
-                    }
                 ],
                 errorElement: <PageNotFound />
             },
@@ -53,7 +46,7 @@ export default createBrowserRouter([
     },
     {
         path: "*",
-        element: <Navigate to="/test" />,
+        element: <Navigate to="/home" />,
         errorElement: <PageNotFound />
     },
 ]);
