--- src/screens/Home/style.scss
@@ -1,15 +1,0 @@
-.mt-3{
-    margin-top: 30px;
-}
-
-.my-3{
-    margin: 30px 0px !important;
-}
-
-.ml-2{
-    margin-left: 20px!important;
-}
-
-.px-3 {
-    padding: 0px 30px;
-}
